Renamed class to be more descriptive. Moved classes.
