--- src/main/java/com/Nxer/TwistSpaceTechnology/common/egs/IEGSBucketFactory.java
@@ -1,14 +1,0 @@
-package com.Nxer.TwistSpaceTechnology.common.egs;
-
-import net.minecraft.nbt.NBTTagCompound;
-
-import com.Nxer.TwistSpaceTechnology.common.machine.TST_MegaTreeFarm;
-
-public interface IEGSBucketFactory {
-
-    String getNBTIdentifier();
-
-    EGSBucket tryCreateBucket(TST_MegaTreeFarm Greenhouse);
-
-    EGSBucket restore(NBTTagCompound nbt);
-}
